requirements/base.txt
requests
pyyaml
sphinx
sphinx-book-theme
sphinx-autobuild
sphinxcontrib-images
sphinxcontrib-youtube
